refactor: collect with-peer tests in directory
