+++ packages/server-wallet/src/__test__/with-peers/crash-tolerance/crash-tolerance.test.ts
@@ -12,8 +12,8 @@ import {
   participantB,
   peersTeardown,
   peerWallets,
-} from '../../../jest/with-peers-setup-teardown';
-import {crashAndRestart, getChannelResultFor, getPayloadFor, ONE_DAY} from '../test-helpers';
+} from '../../../../jest/with-peers-setup-teardown';
+import {crashAndRestart, getChannelResultFor, getPayloadFor, ONE_DAY} from '../../test-helpers';
 
 let channelId: string;
 jest.setTimeout(10_000);
