+++ packages/server-wallet/src/__test__/with-peers/create-and-close-channel/close-channel-simplified.seqdiag
