+++ packages/server-wallet/src/__test__/with-peers/create-and-close-channel/direct-funding-simplified.seqdiag
