+++ packages/server-wallet/src/__test__/with-peers/create-and-close-channel/fake-funding-simplified.seqdiag
